[refactor] 리뷰 파일 정리 및 관리자 리뷰 화면 수정
