+++ eoullim-Project/src/main/java/org/example/eoullimback/admin/review/AdminApiReviewController.java
@@ -4,10 +4,9 @@ import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.example.eoullimback._common.enums.errors.ErrorCode;
 import org.example.eoullimback._common.error.exception.Exception409;
-import org.example.eoullimback.review.ReviewQueryService;
 import org.example.eoullimback.review.ReviewRepository;
 import org.example.eoullimback.review.ReviewService;
-import org.example.eoullimback.review.AdminReviewListDTO;
+import org.example.eoullimback.review.dto.AdminReviewListResponse;
 import org.example.eoullimback.user_auth.user.User;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,19 +18,18 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class AdminApiReviewController {
 
-    private final ReviewQueryService reviewQueryService;
     private final ReviewService reviewService;
     private final ReviewRepository reviewRepository;
 
     @GetMapping("/api/admin/reviews")
-    public List<AdminReviewListDTO> list(HttpSession session,
-                                         @RequestParam(defaultValue = "") String keyword) {
+    public List<AdminReviewListResponse> list(HttpSession session,
+                                              @RequestParam(defaultValue = "") String keyword) {
 
         User sessionUser = (User) session.getAttribute("sessionUser");
         if (sessionUser == null) throw new Exception409(ErrorCode.USER_NOT_FOUND);
 //        if (!sessionUser.isAdmin()) throw new Exception409(ErrorCode.FORBIDDEN);
 
-        return reviewQueryService.findAll(keyword);
+        return reviewService.findAll(keyword);
     }
 
     @DeleteMapping("/api/admin/reviews/{reviewId}")
