+++ eoullim-Project/src/main/java/org/example/eoullimback/place/PlaceController.java
@@ -3,12 +3,11 @@ package org.example.eoullimback.place;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.example.eoullimback._common.dto.PageResponse;
-import org.example.eoullimback.review.ReviewResponse;
+import org.example.eoullimback.review.dto.ReviewResponse;
 import org.example.eoullimback.review.ReviewService;
 import org.example.eoullimback.room.RoomResponse;
 import org.example.eoullimback.room.RoomService;
 import org.example.eoullimback.user_auth.user.User;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
