+++ eoullim-Project/src/main/java/org/example/eoullimback/review/ReviewService.java
@@ -1,6 +1,10 @@
 package org.example.eoullimback.review;
 
 import jakarta.validation.Valid;
+import org.example.eoullimback.review.dto.AdminReviewListResponse;
+import org.example.eoullimback.review.dto.ReviewListItemResponse;
+import org.example.eoullimback.review.dto.ReviewRequest;
+import org.example.eoullimback.review.dto.ReviewResponse;
 
 import java.util.List;
 
@@ -21,4 +25,8 @@ public interface ReviewService {
     List<ReviewResponse.ListDTO> findLatestReviews(int limit);
 
     void adminDelete(Long adminUserId, Long reviewId);
+
+    public List<ReviewListItemResponse> findList(Long userId, String code);
+
+    List<AdminReviewListResponse> findAll(String keyword);
 }
